refactor: 재사용 컴포넌트를 common 폴더로 이동
