+++ src/components/common/MonsterCard/MonsterCard.tsx
