+++ src/components/common/MonsterCard/css/MonsterCard.module.css
