+++ src/components/common/Registration/Descriptions.tsx
