+++ src/components/common/Registration/Label.tsx
