+++ src/components/common/Registration/Quest.tsx
