+++ src/components/common/Registration/Registration.tsx
