+++ src/components/common/Registration/TimerInfo.tsx
