+++ src/components/common/Registration/css/Descriptions.module.css
