+++ src/components/common/Registration/css/Registration.module.css
