+++ src/components/common/Registration/css/TimerInfo.module.css
